Reformatting blog post
The first part of the blog post is aimed for Python beginners. The example that
demonstrates the for-loop is confusing for the beginners. It is not obvious for
them that `string` object is an immutable sequence. Another problem with this
example is that it uses numbers which are actually strings. When the post
demonstrates how to loop over `numbers` with indexes it becomes even more
confusing.

Therefore, using a new example with a simple list object will hopefully
increase comprehensibility.

The second part that is aimed for the more advanced users is somewhat
reformatted in a way that sentences have been flipped so that the story is more
understandable.